requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
alembic
pydantic
pydantic-settings
aiohttp
certifi
alembic